pyproject.toml
[project]
name = "db-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.34.0",
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-mistralai>=0.2.10",
    "langgraph>=0.4.8",
    "mcp>=1.9.4",
    "pydantic>=2.11.7",
    "pymongo>=4.13.1",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
]
